Rename common to enums
